- Added parallel execution support - Cleaned bit of frontend code - Renamed frontend scenarios
