+++ .circleci/config.yml
@@ -29,7 +29,7 @@ jobs:
       # Then run your tests!
       - run:
           name: Test
-          command: mvn clean verify -P dontUseTheForks
+          command: mvn clean verify
 
       - store_test_results:
           path: target/cucumber-report/
