+++ pom.xml
@@ -11,52 +11,10 @@
     <properties>
         <maven.compiler.source>16</maven.compiler.source>
         <maven.compiler.target>16</maven.compiler.target>
-        <serenity.version>3.2.0</serenity.version>
+        <serenity.version>3.2.1</serenity.version>
         <serenity.cucumber.version>3.2.1</serenity.cucumber.version>
         <cucumber.version>7.2.3</cucumber.version>
-        <parallel.tests>3</parallel.tests>
     </properties>
-
-    <profiles>
-        <profile>
-            <id>dontUseTheForks</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <artifactId>maven-failsafe-plugin</artifactId>
-                        <version>3.0.0-M5</version>
-                        <configuration>
-                            <includes>
-                                <include>**/allTestRunner.java</include>
-                            </includes>
-<!--                            <test>**/SlicedTestRunner1.java</test>-->
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>useTheForks</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <artifactId>maven-failsafe-plugin</artifactId>
-                        <version>3.0.0-M5</version>
-                        <configuration>
-                            <includes>
-                                <include>**/SlicedTestRunner*.java</include>
-                            </includes>
-                            <systemPropertyVariables>
-                                <serenity.fork.count>0${parallel.tests}</serenity.fork.count>
-                                <!--suppress UnresolvedMavenProperty -->
-                                <serenity.fork.number>0${surefire.forkNumber}</serenity.fork.number>
-                            </systemPropertyVariables>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
 
     <build>
         <plugins>
@@ -79,14 +37,16 @@
                 <version>3.0.0-M5</version>
                 <configuration>
                     <includes>
-                        <include>**/SlicedTestRunner*.java</include>
+                        <include>**/allTestRunner.java</include>
                     </includes>
                     <systemPropertyVariables>
-<!--                        <webdriver.base.url>${webdriver.base.url}</webdriver.base.url>-->
+                        <!--suppress UnresolvedMavenProperty -->
+                        <webdriver.base.url>${webdriver.base.url}</webdriver.base.url>
                     </systemPropertyVariables>
-                    <parallel>classes</parallel>
-                    <threadCount>${parallel.tests}</threadCount>
-                    <forkCount>${parallel.tests}</forkCount>
+                    <parallel>classesAndMethods</parallel>
+                    <useUnlimitedThreads>true</useUnlimitedThreads>
+                    <threadCount>8</threadCount>
+                    <threadCountClasses>3</threadCountClasses>
                 </configuration>
                 <executions>
                     <execution>
@@ -106,7 +66,6 @@
             <groupId>net.serenity-bdd</groupId>
             <artifactId>serenity-core</artifactId>
             <version>${serenity.version}</version>
-            <scope>test</scope>
             <exclusions>
                 <exclusion>
                     <groupId>io.cucumber</groupId>
@@ -162,7 +121,7 @@
         <dependency>
             <groupId>net.serenity-bdd</groupId>
             <artifactId>serenity-rest-assured</artifactId>
-            <version>3.2.1-RC1</version>
+            <version>${serenity.version}</version>
         </dependency>
 
         <dependency>
@@ -174,7 +133,7 @@
         <dependency>
             <groupId>org.apache.poi</groupId>
             <artifactId>poi-ooxml</artifactId>
-            <version>5.2.1</version>
+            <version>5.2.2</version>
         </dependency>
 
         <dependency>
