+++ src/test/java/frontend/steps/GoogleSteps.java
@@ -2,7 +2,6 @@ package frontend.steps;
 
 import frontend.pages.GooglePages;
 import frontend.pages.IMDBPages;
-import io.cucumber.java.bs.A;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
