+++ src/test/java/runner/allTestRunner.java
@@ -1,5 +1,4 @@
-package runner.parallelRunners;
-
+package runner;
 
 import io.cucumber.junit.CucumberOptions;
 import net.serenitybdd.cucumber.CucumberWithSerenity;
@@ -7,13 +6,12 @@ import org.junit.runner.RunWith;
 
 @RunWith(CucumberWithSerenity.class)
 @CucumberOptions(
-        features = {"src/test/resources/features/backend"},
-        glue = {"backend"},
-        tags = "@test and @backend and not @ignore",
-        stepNotifications = false,
+        features = {"src/test/resources/features"},
+        glue = {""},
+        tags = "@test and not @ignore",
         plugin = {
                 "json:target/cucumber-report/cucumber.json"
         }
 )
-public class SlicedTestRunner3 {
+public class allTestRunner {
 }
