--- src/test/java/runner/parallelRunners/allTestRunner.java
@@ -1,17 +1,0 @@
-package runner.parallelRunners;
-
-import io.cucumber.junit.CucumberOptions;
-import net.serenitybdd.cucumber.CucumberWithSerenity;
-import org.junit.runner.RunWith;
-
-@RunWith(CucumberWithSerenity.class)
-@CucumberOptions(
-        features = {"src/test/resources/features"},
-        glue = {""},
-        tags = "@test and not @ignore",
-        plugin = {
-                "json:target/cucumber-report/cucumber.json"
-        }
-)
-public class allTestRunner {
-}
